Projectiles and dealing damage
Created a health component and attached it to necessary prefabs.

Defined behavior for attackers dealing damage and defenders dying.

Separated projectiles from cactus and gnome prefabs.

Worked out the instantiation of projectiles and created a Shooter component. Wired the firing behavior to the animator.

Added a shredder to remove projectiles from the scene.