Firing States & Projectiles
Added a new state to the gnome's animator so that we can better control the rate at which the fire() method is called in the animation event system.

Added a tag to the "attacking" states for both the Fox and the Lizard so that we can check to see if the state of the animator is in an "attacking" state before we translate the gameObject during the Update().  This fixes a bug where the attacker would move despite being in a attacking animation.

Got projectiles dealing damage to attackers.