Refactorpalooza
Created the GameManager class which contains much of the logic which "ties" together the other various classes.  This simplified the other classes like PinCount, PinSetter, etc.

Created the ScoreMaster.cs class. That was a bear. I only looked at Ben's solution a little (just a peek!). For the record, bowling has weird rules.

Also put in some guardian code to prevent the player (sort of) from launching the ball before the lane was ready.