Goodbye for now
Reached the end of the Udemy course's section on Bowlmaster.  The core gameplay is there, but it is in dire need of some polish. My current UI is abysmal. It doesn't scale appropriately for all devices and the overall aesthetics leave a lot to be desired. Additionally, once the 10 frame is complete the game does nothing, and the controls are a little wonky.

That being said, the game plays reasonably well and is pretty enjoyable. Maybe one day in the future I'll come back and put some polish on this thing.